aider: Converted the project to use TypeScript, added necessary dependencies and plugins, and configured TypeScript options.
# Aider chat conversation:

ASSISTANT: To convert the project to use TypeScript, we need to make several changes to the `package.json` file. We will add TypeScript as a dependency, include the necessary TypeScript plugins for Vite, and add a `tsconfig.json` file to configure TypeScript options.

Here are the diffs for those changes:

```diff
--- package.json
+++ package.json
@@ ... @@
     "devDependencies": {
         "@rushstack/eslint-patch": "^1.6.1",
         "@vitejs/plugin-vue": "^5.0.2",
+        "@vitejs/plugin-vue-jsx": "^1.3.0",
         "@vue/eslint-config-prettier": "^9.0.0",
         "eslint": "^8.56.0",
         "eslint-plugin-vue": "^9.19.2",
         "prettier": "^3.1.1",
         "sass": "^1.69.7",
-        "typescript": "^5.3.3",
+        "typescript": "^4.5.5",
+        "vite-tsconfig-paths": "^3.3.17",
         "vite": "^5.0.10"
     }
 }
```

Additionally, we need to create a `tsconfig.json` file to configure TypeScript for the project:

```diff
--- /dev/null
+++ tsconfig.json
@@ ... @@
+{
+    "compilerOptions": {
+        "target": "esnext",
+        "module": "esnext",
+        "moduleResolution": "node",
+        "strict": true,
+        "jsx": "preserve",
+        "importHelpers": true,
+        "esModuleInterop": true,
+        "allowSyntheticDefaultImports": true,
+        "sourceMap": true,
+        "baseUrl": ".",
+        "paths": {
+            "@/*": ["src/*"]
+        },
+        "lib": ["esnext", "dom"],
+        "types": ["vite/client"],
+        "skipLibCheck": true
+    },
+    "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue"],
+    "exclude": ["node_modules"]
+}
```

Please note that the version of TypeScript has been changed to `^4.5.5` which is a stable version that is compatible with the Vite and Vue ecosystem. The `@vitejs/plugin-vue-jsx` and `vite-tsconfig-paths` plugins have been added to support JSX in Vue components and to allow path mapping in TypeScript, respectively.

After applying these changes, you will need to rename `.js` files to `.ts` and `.vue` files' script tags to `<script lang="ts">`. You may also need to make type annotations and modifications to the code to satisfy TypeScript's type checking.